+++ core/dbt/auth/session_cache.py
@@ -5,7 +5,7 @@ from dataclasses import asdict, dataclass, field
 from pathlib import Path
 
 from dbt.auth.credentials import OAuthSession
-from dbt.auth.utils import secure_open
+from dbt.auth.secure_file import secure_open
 from dbt.exceptions import InaccessibleSource, MalformedAuthConfig
